Restore Gradle wrapper version
