+++ src/main/kotlin/org/example/ContainerFactory.kt
@@ -1,0 +1,5 @@
+package org.example
+
+fun interface ContainerFactory {
+    fun create(route: RouteConfig): Container
+}
